MATCH-66: Cannot match entire/single value
Given a simple pattern like the following:

(match 3 x x)

Bad match code would be generated. The issue was that 3 is an
expression and needs an occurence (a local). However the match
provided by the user here is just a wildcard line. Wildcard line cases
were not checking to see if there are are any expressions in the
occurrences - they just emitted a leaf node, we now check for
expression occurences and emit a wrapping bind node if necessary.